Unify response handlers and split capacity helpers
