+++ src/features/admin/owner-route.ts
@@ -3,6 +3,17 @@ import { requireOwnerOr } from "#routes/auth.ts";
 import { notFoundResponse } from "#routes/response.ts";
 import type { ResponseHandler } from "#shared/response-steps.ts";
 
+/** Owner-gate a request and load the data its handler needs. */
+export const withOwnerData =
+  <T>(load: (session: AuthSession) => Promise<T>) =>
+  (
+    request: Request,
+    handle: ResponseHandler<[session: AuthSession, data: T]>,
+  ): Promise<Response> =>
+    requireOwnerOr(request, async (session) =>
+      handle(session, await load(session)),
+    );
+
 /**
  * Owner-gate a request, look a value up, and 404 when it is absent — otherwise
  * hand the found value (and the session) to `whenFound`. Both the built-sites
@@ -14,7 +25,6 @@ export const ownerFoundOr404 = <T>(
   find: (session: AuthSession) => Promise<T | null>,
   whenFound: ResponseHandler<[found: T, session: AuthSession]>,
 ): Promise<Response> =>
-  requireOwnerOr(request, async (session) => {
-    const found = await find(session);
-    return found === null ? notFoundResponse() : whenFound(found, session);
-  });
+  withOwnerData(find)(request, (session, found) =>
+    found === null ? notFoundResponse() : whenFound(found, session),
+  );
