+++ src/features/public/discovery.ts
@@ -27,7 +27,7 @@ import { getBookableStartDates } from "#shared/dates.ts";
 import {
   getGroupRemainingByListingId,
   getSharedGroupCapacities,
-} from "#shared/db/attendees/capacity.ts";
+} from "#shared/db/attendees/capacity/groups.ts";
 import {
   getActiveListingsByGroupId,
   getActiveListingsByGroupIds,
