+++ src/features/public/order.ts
@@ -40,10 +40,8 @@ import {
 import { createRouter, defineRoutes } from "#routes/router.ts";
 import type { TicketListing } from "#shared/booking/model.ts";
 import { getBookableStartDates } from "#shared/dates.ts";
-import {
-  getGroupRemainingForSpan,
-  getListingRemainingForRange,
-} from "#shared/db/attendees/capacity.ts";
+import { getGroupRemainingForSpan } from "#shared/db/attendees/capacity/groups.ts";
+import { getListingRemainingForRange } from "#shared/db/attendees/capacity/remaining.ts";
 import { getSelectedAttributesForListings } from "#shared/db/attributes.ts";
 import {
   getGroupIdsByListingIds,
