--- src/shared/db/attendees/capacity.ts
@@ -1,754 +1,0 @@
-/**
- * Capacity checks and availability queries for attendees/listing_attendees.
- *
- * Multi-day daily bookings are enforced via per-day expansion: every day in
- * `[start, start + duration_days)` must independently pass listing + group caps.
- * This file contains the JS preflight (`checkListingAvailability`,
- * `checkBatchAvailabilityImpl`) — the inline SQL safety net lives in
- * `#shared/db/capacity.ts` and runs in the same statement as the INSERT/UPDATE.
- * Both sides decide per-date vs running-total counting from the same
- * `#shared/capacity-rules.ts` declaration, so they can never disagree about
- * which check applies to a listing.
- */
-
-import type { InValue } from "@libsql/client";
-import { byId, filter, groupBy, map, mapBy, pipe, sumOf, unique } from "#fp";
-import {
-  capacityDateFor,
-  capacityRuleTypeSql,
-  countsPerDate,
-} from "#shared/capacity-rules.ts";
-import { addDays } from "#shared/dates.ts";
-import type {
-  BatchAvailabilityItem,
-  LineBooking,
-  ListingBooking,
-} from "#shared/db/attendee-types.ts";
-import {
-  buildBatchCapacitySql,
-  buildCapacityCondition,
-  type CapacityBucket,
-  dateToRange,
-} from "#shared/db/capacity.ts";
-import { inPlaceholders, queryAll, queryOne } from "#shared/db/client.ts";
-import { getGroupIdsByListingIds } from "#shared/db/groups.ts";
-import { getListingWithCount } from "#shared/db/listings/records.ts";
-import { columnMapByIds } from "#shared/db/query.ts";
-import {
-  type ListingType,
-  type ListingWithCount,
-  normalizeDurationDays,
-} from "#shared/types.ts";
-
-/** Convert nullable date to start_at/end_at (null-safe wrapper around dateToRange) */
-export const dateToStartEnd = (
-  date: string | null,
-  durationDays = 1,
-): { startAt: string | null; endAt: string | null } => {
-  if (!date) return { endAt: null, startAt: null };
-  const range = dateToRange(date, durationDays);
-  return { endAt: range.endAt, startAt: range.startAt };
-};
-
-type RemainingMap = Map<number, number>;
-
-type RemainingLookup<Input> = (
-  inputs: Input[],
-  date?: string | null,
-) => Promise<RemainingMap>;
-
-/** Distinct group ids worth a cap lookup — positive only (0 = ungrouped). */
-const uniquePositiveGroupIds = (groupIds: number[]): number[] =>
-  unique(groupIds.filter((id) => id > 0));
-
-/** Run `queryFor` over the distinct capped-group candidate ids, or short-circuit
- * to an empty map when none are worth a lookup. The shared dedup + empty guard
- * every per-group cap query opens with. */
-const cappedGroupQuery = async <T>(
-  groupIds: number[],
-  queryFor: (ids: number[]) => Promise<Map<number, T>>,
-): Promise<Map<number, T>> => {
-  const ids = uniquePositiveGroupIds(groupIds);
-  return ids.length === 0 ? new Map() : queryFor(ids);
-};
-
-/**
- * Per-group remaining capacity. Groups with `max_attendees <= 0` (no cap)
- * are omitted from the map. With `date = null`, this uses each listing's
- * editable booked_quantity running total; with a date, non-daily listings still
- * use booked_quantity while daily listings count overlapping attendee rows.
- */
-export const getGroupRemainingByGroupId: RemainingLookup<number> = (
-  groupIds,
-  date = null,
-) =>
-  cappedGroupQuery(groupIds, async (ids) => {
-    const range = date ? dateToRange(date) : null;
-    const datedCount = date
-      ? `COALESCE((
-        SELECT SUM(listing.booked_quantity)
-          FROM listings AS listing
-          JOIN group_listings AS groupListing ON groupListing.listing_id = listing.id
-         WHERE groupListing.group_id = groupRow.id AND ${capacityRuleTypeSql(
-           "dateLessCap",
-           "listing.listing_type",
-         )}
-      ), 0) + COALESCE((
-        SELECT SUM(attendee.quantity)
-          FROM listing_attendees AS attendee
-          JOIN listings AS listing ON listing.id = attendee.listing_id
-          JOIN group_listings AS groupListing ON groupListing.listing_id = attendee.listing_id
-         WHERE groupListing.group_id = groupRow.id
-           AND ${capacityRuleTypeSql("perDateCap", "listing.listing_type")}
-           AND attendee.start_at < ? AND attendee.end_at > ?
-      ), 0)`
-      : `COALESCE((
-        SELECT SUM(listing.booked_quantity)
-          FROM listings AS listing
-          JOIN group_listings AS groupListing ON groupListing.listing_id = listing.id
-         WHERE groupListing.group_id = groupRow.id
-      ), 0)`;
-    const countArgs = range ? [range.endAt, range.startAt] : [];
-    const rows = await queryAll<{
-      group_id: number;
-      max_attendees: number;
-      count: number;
-    }>(
-      `SELECT groupRow.id as group_id, groupRow.max_attendees,
-            ${datedCount} as count
-     FROM groups AS groupRow
-     WHERE groupRow.id IN (${inPlaceholders(
-       ids,
-     )}) AND groupRow.max_attendees > 0
-     GROUP BY groupRow.id`,
-      [...countArgs, ...ids],
-    );
-    return mapBy("group_id", (row: (typeof rows)[number]) =>
-      Math.max(0, row.max_attendees - row.count),
-    )(rows);
-  });
-
-/** Per-group remaining over a whole daily span: the tightest day in
- * `[date, date + spanDays)`, because a booking must fit EVERY day it
- * occupies. With no date this is the plain running-total remaining, and a
- * one-day span matches the single-date figure. Uncapped groups are omitted. */
-export const getGroupRemainingForSpan = async (
-  groupIds: number[],
-  date: string | null,
-  spanDays = 1,
-): Promise<RemainingMap> => {
-  if (date === null) return getGroupRemainingByGroupId(groupIds, null);
-  const days = expandDailyRange(date, spanDays);
-  const perDay = await groupPerDayRemainingByGroup(groupIds, days);
-  return new Map(
-    [...perDay].map(([groupId, byDay]) => [
-      groupId,
-      Math.min(...days.map((day) => byDay.get(day)!)),
-    ]),
-  );
-};
-
-/** Date-less remaining for every capped group reachable from a `dateLessCap`
- * listing in `members`. A `perDateCap` listing's group remaining is a
- * per-DATE fact — a date-less cumulative count would misreport spots other
- * dates still have — so those members contribute no date-less pool clamp,
- * exactly like the standalone paths; the authoritative per-date group check
- * runs at booking. Membership maps stay complete for demand computation; only
- * the remaining lookup is filtered. */
-export const getDatelessGroupRemaining = (
-  members: readonly { id: number; listing_type: ListingType }[],
-  membership: ReadonlyMap<number, number[]>,
-): Promise<RemainingMap> =>
-  getGroupRemainingByGroupId([
-    ...new Set(
-      members
-        .filter((m) => !countsPerDate(m.listing_type))
-        .flatMap((m) => membership.get(m.id) ?? []),
-    ),
-  ]);
-
-type ListingForGroupLookup = {
-  id: number;
-  listing_type: ListingType;
-};
-
-/**
- * For each listing, the tightest (minimum) value across the groups it belongs
- * to that appear in `byGroup` (the capped groups). A listing in several capped
- * groups is constrained by whichever group has the least headroom; listings
- * with no capped group are omitted, matching the old single-group behaviour for
- * ungrouped/uncapped listings.
- */
-const minByListingOverGroups = (
-  listingIds: number[],
-  membership: Map<number, number[]>,
-  byGroup: RemainingMap,
-): RemainingMap => {
-  const result: RemainingMap = new Map();
-  for (const id of listingIds) {
-    const values = (membership.get(id) ?? [])
-      .map((g) => byGroup.get(g))
-      .filter((v): v is number => v !== undefined);
-    if (values.length > 0) result.set(id, Math.min(...values));
-  }
-  return result;
-};
-
-/** Load the listings' group membership along with the flat list of every group
- * id any of them belongs to — the pair every per-listing → per-group lookup
- * below starts from. */
-const loadMembershipWithGroupIds = async (
-  listings: ListingForGroupLookup[],
-): Promise<{ membership: Map<number, number[]>; groupIds: number[] }> => {
-  const membership = await getGroupIdsByListingIds(listings.map((e) => e.id));
-  return { groupIds: [...membership.values()].flat(), membership };
-};
-
-/**
- * Per-listing view of group remaining capacity. `perDateCap` listings are
- * dropped when `date` is null — their cap is per-date, so a cumulative count
- * would misreport spots that other dates still have. A listing in multiple
- * capped groups reports the tightest group's remaining.
- */
-export const getGroupRemainingByListingId: RemainingLookup<
-  ListingForGroupLookup
-> = async (listings, date = null) => {
-  const candidates = date
-    ? listings
-    : listings.filter((e) => !countsPerDate(e.listing_type));
-  const { membership, groupIds } = await loadMembershipWithGroupIds(candidates);
-  const groupMap = await getGroupRemainingByGroupId(groupIds, date);
-  return minByListingOverGroups(
-    candidates.map((e) => e.id),
-    membership,
-    groupMap,
-  );
-};
-
-/**
- * Per-GROUP STATIC cap (`groups.max_attendees`), date-INDEPENDENT, keyed by group
- * id. Uncapped/absent groups are omitted, matching {@link getGroupRemainingByGroupId}.
- *
- * The static cap is a structural fact (how many can EVER sit in the group), not a
- * per-date count, so date-less surfaces use it to reject a parent+child share that
- * can never satisfy the combined minimum order (a parent and its required child
- * co-grouped consume `PARENT_CHILD_GROUP_UNITS` spots) even when a daily child's
- * per-date remaining is unknown without a date.
- */
-export const getGroupStaticCapByGroupId = (
-  groupIds: number[],
-): Promise<RemainingMap> =>
-  cappedGroupQuery(groupIds, (ids) =>
-    columnMapByIds(
-      "groups",
-      "groupRow",
-      "max_attendees",
-      ids,
-      " AND groupRow.max_attendees > 0",
-    ),
-  );
-
-/** The PER-GROUP shared-group capacity facts + membership for a set of listings,
- * the single fetch every date-less parent/child surface (discovery + the booking
- * page's sold-out projection + the quantity clamp) uses. Returns the date-less
- * `remaining` and date-independent `staticCap` keyed by GROUP id (not per
- * listing), plus each listing's group `membership`, so the {@link
- * SharedGroupCapacity} vocabulary can reason about the SPECIFIC group a parent and
- * child share rather than a child's tightest group overall. */
-export const getSharedGroupCapacities = async (
-  listings: ListingForGroupLookup[],
-): Promise<{
-  remaining: RemainingMap;
-  staticCap: RemainingMap;
-  membership: Map<number, number[]>;
-}> => {
-  const { membership, groupIds: allGroupIds } =
-    await loadMembershipWithGroupIds(listings);
-  // A daily listing's group `remaining` is a per-DATE fact (see
-  // getDatelessGroupRemaining). The structural `staticCap` is date-independent,
-  // so it covers every group.
-  const [remaining, staticCap] = await Promise.all([
-    getDatelessGroupRemaining(listings, membership),
-    getGroupStaticCapByGroupId(allGroupIds),
-  ]);
-  return { membership, remaining, staticCap };
-};
-
-/** Resolve a bare listing id to its row and read that listing's entry from a
- * map-producing remaining lookup — `undefined` when the listing doesn't exist.
- * The shared singular path behind the array-or-id lookups below (a single
- * listing is just an array of one). */
-const useListingById = async <Result>(
-  listingId: number,
-  missing: Result,
-  useListing: (listing: ListingWithCount) => Promise<Result>,
-): Promise<Result> => {
-  const listing = await getListingWithCount(listingId);
-  return listing ? useListing(listing) : missing;
-};
-
-/** Returns `undefined` when no group cap applies: ungrouped, uncapped
- * group, or daily listing without a `date`. */
-export async function getGroupRemainingForListing(
-  listingOrId: ListingForGroupLookup | number,
-  date: string | null = null,
-): Promise<number | undefined> {
-  const lookup = (listing: ListingForGroupLookup): Promise<RemainingMap> =>
-    getGroupRemainingByListingId([listing], date);
-  return typeof listingOrId === "number"
-    ? useListingById(listingOrId, undefined, async (listing) =>
-        (await lookup(listing)).get(listingOrId),
-      )
-    : (await lookup(listingOrId)).get(listingOrId.id);
-}
-
-/**
- * Build an INSERT into listing_attendees, capacity-checked by default.
- * @param attendeeIdExpr - SQL expression for attendee_id (e.g. "last_insert_rowid()" or "?")
- * @param attendeeIdArg - Argument for "?" expr, omit for last_insert_rowid()
- * @param allowOverbook - when true the capacity WHERE is dropped so the row is
- *   inserted unconditionally (admin manual add/edit may deliberately overbook).
- */
-export const buildCapacityCheckedInsert = (
-  booking: ListingBooking,
-  attendeeIdExpr = "last_insert_rowid()",
-  attendeeIdArg?: number,
-  allowOverbook = false,
-): { sql: string; args: InValue[] } => {
-  const {
-    listingId,
-    quantity: qty = 1,
-    date = null,
-    durationDays = 1,
-    orderToken = "",
-    parentListingId = 0,
-    packageGroupId = 0,
-  } = booking;
-  const { startAt, endAt } = dateToStartEnd(date, durationDays);
-  const args: InValue[] = [listingId];
-  if (attendeeIdArg !== undefined) args.push(attendeeIdArg);
-  args.push(startAt, endAt, qty, orderToken, parentListingId, packageGroupId);
-  // price_paid is no longer stored — a booking row's amount paid projects from
-  // its ledger sale leg (posted by the booking poster from booking.pricePaid).
-  // The order token + parent listing still persist for the parent/child gate;
-  // package_group_id groups the order's lines under the package on tickets/emails.
-  const insertSelect = `INSERT INTO listing_attendees (listing_id, attendee_id, start_at, end_at, quantity, order_token, parent_listing_id, package_group_id)
-          SELECT ?, ${attendeeIdExpr}, ?, ?, ?, ?, ?, ?`;
-  if (allowOverbook) return { args, sql: insertSelect };
-
-  const condition = buildCapacityCondition(
-    listingId,
-    qty,
-    date,
-    undefined,
-    durationDays,
-  );
-  args.push(...condition.args);
-  return { args, sql: `${insertSelect}\n          WHERE ${condition.sql}` };
-};
-
-// ---------------------------------------------------------------------------
-// Per-day load helpers for the remaining-capacity lookup and batch preflight.
-//
-// Per-day loads are computed by fetching every row overlapping the whole
-// booking span ONCE and summing per-day in JS — one round trip instead of
-// one per day, which matters because production talks to a remote libsql
-// server where each query is a network hop.
-// ---------------------------------------------------------------------------
-
-/** Expand a daily-listing range into individual day strings. */
-export const expandDailyRange = (
-  date: string,
-  durationDays: number,
-): string[] => {
-  const duration = normalizeDurationDays(durationDays);
-  return Array.from({ length: duration }, (_, i) => addDays(date, i));
-};
-
-/** Half-open [startAt, endAt) span covering a set of YYYY-MM-DD days. */
-const daySpan = (days: string[]): { startAt: string; endAt: string } => {
-  const sorted = days.toSorted();
-  return {
-    endAt: dateToRange(sorted.at(-1)!).endAt,
-    startAt: `${sorted[0]!}T00:00:00Z`,
-  };
-};
-
-/** A booking row's range + quantity. Ranges are never null here — the
- * fetch queries below only match rows via the overlap predicate, which
- * NULL ranges can never satisfy (mirroring the SQL safety net). */
-type IntervalRow = { start_at: string; end_at: string; quantity: number };
-
-/** Sum the quantity column of a set of rows */
-const sumQuantity = sumOf((row: { quantity: number }) => row.quantity);
-
-/** Curried day-overlap predicate. String comparison mirrors SQLite TEXT
- * comparison byte-for-byte, so this reproduces the SQL overlap predicate
- * `start_at < dayEnd AND end_at > dayStart` exactly. */
-export const overlapsDay = (day: string) => {
-  const { startAt, endAt } = dateToRange(day);
-  return (row: IntervalRow): boolean =>
-    row.start_at < endAt && row.end_at > startAt;
-};
-
-/** Per-day quantity sums for the given days from pre-fetched rows. */
-const perDayLoads = (
-  rows: IntervalRow[],
-  days: string[],
-): Map<string, number> =>
-  new Map(
-    map((day: string): [string, number] => [
-      day,
-      pipe(filter(overlapsDay(day)), sumQuantity)(rows),
-    ])(days),
-  );
-
-/**
- * Per-day availability check for a single-listing booking. Used by
- * `hasAvailableSpots` (public availability API).
- *
- * The capacity rules — per-day for daily listings, cumulative total for
- * standard/date-less, plus group caps — live entirely in
- * `buildCapacityCondition`, the same SQL the atomic write uses, so the whole
- * check is one query (the listing lookup is a cache hit on the booking paths).
- * Does NOT self-exclude — admin edit paths rely on the atomic SQL guard instead.
- */
-export const checkListingAvailability = async (
-  listingId: number,
-  quantity = 1,
-  date?: string | null,
-  durationDays = 1,
-): Promise<boolean> => {
-  return useListingById(listingId, false, async (listing) => {
-    // A dateLessCap listing's rows carry no booking range, so they'd never match
-    // a date overlap — capacityDateFor drops the date to count the running total.
-    const checkDate = capacityDateFor(listing.listing_type, date);
-    return (
-      await checkLinesCapacity([
-        { date: checkDate, durationDays, listingId, quantity },
-      ])
-    )[0]!;
-  });
-};
-
-type ListingRow = {
-  id: number;
-  max_attendees: number;
-  listing_type: ListingType;
-  attendee_count: number;
-};
-
-/** The demand split a listing/group bucket carries — identical to the SQL
- * builder's {@link CapacityBucket}, so aggregated demand feeds it directly. */
-type DemandBucket = CapacityBucket;
-
-/** Get the bucket for `key`, creating an empty one on first use. */
-const getOrCreateBucket = <K>(
-  buckets: Map<K, DemandBucket>,
-  key: K,
-): DemandBucket => {
-  let bucket = buckets.get(key);
-  if (!bucket) {
-    bucket = { perDay: new Map(), total: 0 };
-    buckets.set(key, bucket);
-  }
-  return bucket;
-};
-
-/** Add one item's demand to a bucket: per-day for a dated `perDateCap`
- * listing, otherwise a single total. */
-const addDemandToBucket = (
-  bucket: DemandBucket,
-  ev: ListingRow,
-  item: BatchAvailabilityItem,
-  date: string | null | undefined,
-): void => {
-  if (countsPerDate(ev.listing_type) && date) {
-    for (const day of expandDailyRange(date, item.durationDays ?? 1)) {
-      bucket.perDay.set(day, (bucket.perDay.get(day) ?? 0) + item.quantity);
-    }
-  } else {
-    bucket.total += item.quantity;
-  }
-};
-
-/**
- * Aggregate batch items into demand buckets keyed by whatever `keysFor` returns
- * for each item. Keyed by `[ev.id]` it gives per-listing demand (listing caps);
- * keyed by the listing's group ids it gives per-group demand (group caps) — a
- * listing in several groups contributes to each, so each group's cap is checked.
- */
-const aggregateDemand = (
-  ctx: BatchAvailabilityContext,
-  keysFor: (ev: ListingRow, item: BatchAvailabilityItem) => number[],
-): Map<number, DemandBucket> => {
-  const { items, listingsById, date } = ctx;
-  const buckets = new Map<number, DemandBucket>();
-  for (const item of items) {
-    const ev = listingsById.get(item.listingId)!;
-    for (const key of keysFor(ev, item)) {
-      addDemandToBucket(getOrCreateBucket(buckets, key), ev, item, date);
-    }
-  }
-  return buckets;
-};
-
-/**
- * Shared inputs to a batch-capacity check: the items under test, the resolved
- * listing rows they reference, and the optional anchor date for daily listings.
- */
-type BatchAvailabilityContext = {
-  items: BatchAvailabilityItem[];
-  listingsById: Map<number, ListingRow>;
-  date: string | null | undefined;
-};
-
-/**
- * Check availability for multiple listings in a single preflight pass.
- *
- * Aggregates the cart's combined demand per listing and per group (per-day for
- * dated daily listings), then evaluates it with ONE SELECT built from the SAME
- * count subqueries the atomic write predicate uses ({@link
- * buildBatchCapacitySql}). Reusing the write's counting is what keeps this
- * read-time preflight and the write-time guard from ever disagreeing about
- * capacity — a single round trip, no per-listing or per-group fan-out.
- */
-export const checkBatchAvailabilityImpl = async (
-  items: BatchAvailabilityItem[],
-  date?: string | null,
-): Promise<boolean> => {
-  if (items.length === 0) return true;
-  // Reject negative quantities outright — would otherwise offset positive
-  // rows and bypass the cap. Form validation clamps upstream; defensive.
-  if (items.some((i) => i.quantity < 0)) return false;
-  const listingIds = map((i: BatchAvailabilityItem) => i.listingId)(items);
-
-  // The listing_type per item drives per-day vs total demand bucketing; the
-  // cap comparison itself is left to the SQL (which also enforces active = 1).
-  const listingRows = await queryAll<ListingRow>(
-    `SELECT listing.id, listing.max_attendees, listing.listing_type,
-            listing.booked_quantity as attendee_count
-     FROM listings AS listing
-     WHERE listing.id IN (${inPlaceholders(listingIds)})`,
-    listingIds,
-  );
-  const listingsById = byId(listingRows);
-
-  // Every item must reference a known listing.
-  if (items.some((i) => !listingsById.has(i.listingId))) return false;
-
-  const membership = await getGroupIdsByListingIds(listingIds);
-  const ctx: BatchAvailabilityContext = { date, items, listingsById };
-  const listingDemand = aggregateDemand(ctx, (ev) => [ev.id]);
-  const groupDemand = aggregateDemand(
-    ctx,
-    (_ev, item) => membership.get(item.listingId) ?? [],
-  );
-
-  const { sql, args } = buildBatchCapacitySql(listingDemand, groupDemand);
-  const row = (await queryOne<{ fits: number }>(sql, args))!;
-  return row.fits === 1;
-};
-
-// ---------------------------------------------------------------------------
-// Remaining-capacity lookup (display + overbooking warnings)
-// ---------------------------------------------------------------------------
-
-/** A listing's identity + the capacity inputs a remaining-units lookup needs —
- * the same shape the batch check uses. `ListingWithCount` satisfies it. */
-export type ListingCapacityRow = ListingRow;
-
-/** Loads a per-id map (keyed by listing or group id) from an id list and the
- * days to cover — the shape the overlapping-rows and per-day-remaining lookups
- * both have. */
-type PerIdDayLoader<Value> = (
-  ids: number[],
-  days: string[],
-) => Promise<Map<number, Value>>;
-
-/** All overlapping interval rows for several listings in one query, grouped by
- * listing id — so per-day loads come from one round trip, not one per listing. */
-const overlappingRowsByListing: PerIdDayLoader<IntervalRow[]> = async (
-  listingIds,
-  days,
-) => {
-  if (listingIds.length === 0) return new Map();
-  const { startAt, endAt } = daySpan(days);
-  const rows = await queryAll<IntervalRow & { listing_id: number }>(
-    `SELECT listing_id, start_at, end_at, quantity FROM listing_attendees
-     WHERE listing_id IN (${inPlaceholders(
-       listingIds,
-     )}) AND start_at < ? AND end_at > ?`,
-    [...listingIds, endAt, startAt],
-  );
-  return groupBy(rows, (row) => row.listing_id);
-};
-
-/** Per-day group remaining for several groups in two queries (caps + occupancy
- * rows), keyed by group id. Uncapped/absent groups are omitted, matching
- * `getGroupRemainingByGroupId`. */
-const groupPerDayRemainingByGroup: PerIdDayLoader<Map<string, number>> = async (
-  groupIds,
-  days,
-) => {
-  const result = new Map<number, Map<string, number>>();
-  const ids = uniquePositiveGroupIds(groupIds);
-  if (ids.length === 0) return result;
-  const caps = await queryAll<{
-    id: number;
-    max_attendees: number;
-    base: number;
-  }>(
-    `SELECT groupRow.id, groupRow.max_attendees,
-            COALESCE((
-              SELECT SUM(listing.booked_quantity)
-                FROM listings AS listing
-                JOIN group_listings AS groupListing ON groupListing.listing_id = listing.id
-               WHERE groupListing.group_id = groupRow.id AND ${capacityRuleTypeSql(
-                 "dateLessCap",
-                 "listing.listing_type",
-               )}
-            ), 0) AS base
-       FROM groups AS groupRow
-     WHERE groupRow.id IN (${inPlaceholders(
-       ids,
-     )}) AND groupRow.max_attendees > 0`,
-    ids,
-  );
-  if (caps.length === 0) return result;
-  const cappedIds = caps.map((c) => c.id);
-  const { startAt, endAt } = daySpan(days);
-  type GroupRow = IntervalRow & { group_id: number };
-  const rows = await queryAll<GroupRow>(
-    `SELECT groupListing.group_id, attendee.start_at, attendee.end_at, attendee.quantity
-     FROM listing_attendees AS attendee
-     JOIN listings AS listing ON listing.id = attendee.listing_id
-     JOIN group_listings AS groupListing ON groupListing.listing_id = attendee.listing_id
-     WHERE groupListing.group_id IN (${inPlaceholders(cappedIds)})
-       AND ${capacityRuleTypeSql("perDateCap", "listing.listing_type")}
-       AND attendee.start_at < ? AND attendee.end_at > ?`,
-    [...cappedIds, endAt, startAt],
-  );
-  const rowsByGroup = groupBy(rows, (row) => row.group_id);
-  for (const { id, max_attendees, base } of caps) {
-    const groupRows = rowsByGroup.get(id) ?? [];
-    const loads = perDayLoads(groupRows, days);
-    // perDayLoads always has an entry for every day in `days`, so loads.get(day)
-    // is never undefined here.
-    result.set(
-      id,
-      new Map(days.map((day) => [day, max_attendees - base - loads.get(day)!])),
-    );
-  }
-  return result;
-};
-
-/**
- * Remaining bookable units per listing for an anchor date + duration, as a
- * `Map<listingId, number>` (clamped at 0).
- *
- * The read-side mirror of the booking-time caps in `checkListingAvailability`:
- * same per-day expansion and group-cap rules, returning the count for display
- * ("X/Y remaining") and overbooking warnings. Daily listings (with a date) are
- * evaluated per day and report the tightest day; standard listings — and every
- * listing when `date` is null — use their cumulative totals.
- *
- * Batched into a constant ≤4 queries regardless of how many listings are passed,
- * so a large catalogue can't blow the per-request query budget.
- */
-export function getListingRemainingForRange(
-  listings: ListingCapacityRow[],
-  date: string | null,
-  durationDays?: number,
-): Promise<Map<number, number>>;
-export function getListingRemainingForRange(
-  listingId: number,
-  date: string | null,
-  durationDays?: number,
-): Promise<number | undefined>;
-export async function getListingRemainingForRange(
-  listingsOrId: ListingCapacityRow[] | number,
-  date: string | null,
-  durationDays = 1,
-): Promise<Map<number, number> | number | undefined> {
-  if (typeof listingsOrId === "number") {
-    return useListingById(listingsOrId, undefined, async (listing) =>
-      (await getListingRemainingForRange([listing], date, durationDays)).get(
-        listingsOrId,
-      ),
-    );
-  }
-  const listings = listingsOrId;
-  const usesRange = (l: ListingCapacityRow): boolean =>
-    countsPerDate(l.listing_type) && date !== null;
-  const daily = filter(usesRange)(listings);
-  const totals = filter((l: ListingCapacityRow) => !usesRange(l))(listings);
-  const days = date ? expandDailyRange(date, durationDays) : [];
-
-  const membership = await getGroupIdsByListingIds(listings.map((l) => l.id));
-  const groupsOf = (l: ListingCapacityRow): number[] =>
-    membership.get(l.id) ?? [];
-
-  const [totalGroupRemaining, overlapByListing, dailyGroupPerDay] =
-    await Promise.all([
-      getGroupRemainingByGroupId(totals.flatMap(groupsOf), null),
-      overlappingRowsByListing(
-        daily.map((l) => l.id),
-        days,
-      ),
-      groupPerDayRemainingByGroup(daily.flatMap(groupsOf), days),
-    ]);
-
-  const totalRemaining = totals.map((l): [number, number] => {
-    const base = l.max_attendees - l.attendee_count;
-    const groupRemainings = groupsOf(l)
-      .map((g) => totalGroupRemaining.get(g))
-      .filter((r): r is number => r !== undefined);
-    return [l.id, Math.min(base, ...groupRemainings)];
-  });
-  const dailyRemaining = daily.map((l): [number, number] => {
-    const loads = perDayLoads(overlapByListing.get(l.id) ?? [], days);
-    // perDayLoads and dailyGroupPerDay both key every day in `days`, so the
-    // loads.get(day)!/m.get(day)! lookups below are never undefined.
-    const listingRemaining = Math.min(
-      ...days.map((day) => l.max_attendees - loads.get(day)!),
-    );
-    const groupPerDayMins = groupsOf(l)
-      .map((g) => dailyGroupPerDay.get(g))
-      .filter((m): m is Map<string, number> => m !== undefined)
-      .map((m) => Math.min(...days.map((day) => m.get(day)!)));
-    return [l.id, Math.min(listingRemaining, ...groupPerDayMins)];
-  });
-  return new Map([...totalRemaining, ...dailyRemaining]);
-}
-
-/**
- * Batched capacity check: whether each booking fits, in a single query. The
- * per-booking self-excluding conditions (the same ones the atomic write uses)
- * become separate columns of one SELECT, so N lines cost one round trip instead
- * of N. Pass `excludeAttendeeId` to ignore that attendee's own rows, so an
- * unchanged edit doesn't count against itself. Drives the admin overbooking
- * warning and the edit preflight; the save itself is allowed to overbook.
- */
-export const checkLinesCapacity = async (
-  bookings: LineBooking[],
-  excludeAttendeeId?: number,
-): Promise<boolean[]> => {
-  if (bookings.length === 0) return [];
-  const conditions = bookings.map((b) =>
-    buildCapacityCondition(
-      b.listingId,
-      b.quantity,
-      b.date,
-      excludeAttendeeId,
-      b.durationDays,
-    ),
-  );
-  const columns = conditions.map((c, i) => `(${c.sql}) AS ok${i}`).join(", ");
-  const args = conditions.flatMap((c) => c.args);
-  const row = (await queryOne<Record<string, number>>(
-    `SELECT ${columns}`,
-    args,
-  ))!;
-  return conditions.map((_, i) => row[`ok${i}`] === 1);
-};
