+++ test/lib/db/attendees/dateless-group-remaining.test.ts
@@ -1,6 +1,6 @@
 import { expect } from "@std/expect";
 import { it as test } from "@std/testing/bdd";
-import { getDatelessGroupRemaining } from "#shared/db/attendees/capacity.ts";
+import { getDatelessGroupRemaining } from "#shared/db/attendees/capacity/groups.ts";
 import { getGroupIdsByListingIds } from "#shared/db/groups.ts";
 import { describeWithEnv } from "#test-utils/db.ts";
 import { createTestGroup } from "#test-utils/db-helpers/groups.ts";
