+++ test/shared/db/attendees/servicing/capacity-blocking.test.ts
@@ -18,11 +18,9 @@
 import { expect } from "@std/expect";
 import { it as test } from "@std/testing/bdd";
 import { hasAvailableSpots } from "#shared/db/attendees/api.ts";
-import {
-  checkListingAvailability,
-  getGroupRemainingForListing,
-  getListingRemainingForRange,
-} from "#shared/db/attendees/capacity.ts";
+import { checkListingAvailability } from "#shared/db/attendees/capacity/checks.ts";
+import { getGroupRemainingForListing } from "#shared/db/attendees/capacity/groups.ts";
+import { getListingRemainingForRange } from "#shared/db/attendees/capacity/remaining.ts";
 import { describeWithEnv } from "#test-utils/db.ts";
 import { createTestGroup } from "#test-utils/db-helpers/groups.ts";
 import {
