+++ test/shared/db/attendees/servicing/duplicate.test.ts
@@ -15,7 +15,7 @@
 // jscpd:ignore-start
 import { expect } from "@std/expect";
 import { describe, it as test } from "@std/testing/bdd";
-import { getListingRemainingForRange } from "#shared/db/attendees/capacity.ts";
+import { getListingRemainingForRange } from "#shared/db/attendees/capacity/remaining.ts";
 import { SERVICING_KIND } from "#shared/db/attendees/kind.ts";
 import { describeWithEnv } from "#test-utils/db.ts";
 import {
